libclamav: drop cl_settempdir(); use cl_engine_set() with CL_ENGINE_TMPDIR and CL_ENGINE_KEEPTMP instead
git-svn: trunk@4416

+++ shared/cfgparser.h
@@ -46,8 +46,8 @@ struct cfgstruct {
     char *optname;
     char *strarg;
     int numarg;
-    short enabled;
-    short multiple;
+    unsigned int enabled;
+    unsigned int multiple;
     struct cfgstruct *nextarg;
     struct cfgstruct *next;
 };
